Update flux.js
Misinformation about COVID-19 will cost people their lives. And right now there’s a ton of false information about coronavirus circulating on the airwaves.

Notoriously racist radio hosts Sean Hannity, Rush Limbaugh and other right-wing broadcasters downplayed the coronavirus in the early weeks of the crisis. We need the FCC to help stop the spread of misinformation no matter where it’s coming from.

The Federal Communications Commission has rules on the books to deal with this kind of misuse and abuse of the public airwaves and has a responsibility to hold broadcasters accountable for what they air.

Sign the petition: Tell the FCC to step in and stop the spread of misinformation about COVID-19 on the public airwaves NOW.